% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testrun_queries.R
\name{adapt_query_to_schema}
\alias{adapt_query_to_schema}
\title{Modify Query to mention the schema that is used}
\usage{
adapt_query_to_schema(query, table_name, schema)
}
\arguments{
\item{table_name}{character table name}

\item{schema}{character schema name}

\item{query:}{character sql query}
}
\value{
query: character modified query
}
\description{
Add schema to the tables if the postgres instance
relates to a schema
}
\examples{
\dontrun{
  adapt_query_to_schema(
    "SELECT COUNT(*) as anzahl_abstimmungsvorlagen FROM abstimmungsvorlagen_seit_1971 as T JOIN spatial_unit as S on T.spatialunit_uid = S.spatialunit_uid;"
    FALSE,
    "abstimmungsvorlagen_seit_1971",
    "experiment"
  )
}
}
