% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a pipeline}
\usage{
create_dataset(id)
}
\arguments{
\item{id}{id of the dataset}
}
\value{
ds_class class containing dataset with attributes
}
\description{
Creates a dataset object for the statbot datasets
}
\details{
A dataset object needs to be created for the download, the computation
as well as the publishing process

A dataset object inherits the following classes:
\itemize{
\item data_organization -> needed for the download process
\item dataset_id -> needed for the download as well as the publishing process
\item download_format -> needed for the download process
}
}
\concept{Datensatz erstellen}
