% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db.R
\name{delete_table_in_postgres}
\alias{delete_table_in_postgres}
\title{Delete table from postgres}
\usage{
delete_table_in_postgres(table_name)
}
\arguments{
\item{table_name}{name of the table to delete}
}
\description{
Delete table from postgres
}
\examples{
\dontrun{
delete_metadata_from_postgres(ds)
}
}
