% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_mappings.R
\name{filter_and_add_pattern}
\alias{filter_and_add_pattern}
\title{Filter and add a pattern out of the names of the spatial unit}
\usage{
filter_and_add_pattern(map_df, spatial_dimensions, pattern_function)
}
\arguments{
\item{spatial_dimensions}{a vector of values for spatialunit_ontology}

\item{pattern_function}{function to make a pattern for detecting
the spatial unit, default stattab_make_pattern
as an example}
}
\value{
spatial_unit_df a tibble with spatial units of the given
spacial_ontologies with a column added for the pattern
}
\description{
Filter the spatial units by their spatial ontology values and
add a pattern from the name and its translations
}
\examples{
\dontrun{
# get a tibble with the spatial units on the levels country and canton
df_spatial <- filter_and_add_pattern(map_df, c("Country", "Canton"))
}
}
