% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{generate_metadata_templates}
\alias{generate_metadata_templates}
\title{generate templates for metadata tables as csv file}
\usage{
generate_metadata_templates(ds)
}
\arguments{
\item{ds}{the dataset: it is expected to have
ds$postgres_export}
}
\description{
The function writes template files for the metadata
of table and column. These template can be manually
completed and copied into the expected metadata files
for the table and columns
}
\examples{
\dontrun{
  generate_metadata_templates(ds)
}
}
