% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closest_matches.R
\name{get_closest_matches}
\alias{get_closest_matches}
\title{Get Closest Matches for Non-matched Spatial Units}
\usage{
get_closest_matches(non_matched_list, spatial_units)
}
\arguments{
\item{non_matched_list}{A named list of character vectors, where each vector contains the names
of non-matched spatial units for a particular spatial unit ontology.
The names of the list elements are the spatial unit ontologies.}

\item{spatial_units}{A data frame containing the spatial units information,
where each row is a spatial unit and there are columns for the name of the spatial unit
in multiple languages (name, name_de, name_fr, name_it).}
}
\value{
A named list where each element is a character vector containing the closest matches
for the non-matched elements in the corresponding list element of \code{non_matched_list}.
The names of the list elements are the spatial unit ontologies.
}
\description{
This function returns the closest matches from a list of non-matched spatial units,
where the matches are sought across multiple name columns of a spatial units data frame.
}
\examples{
\dontrun{
# Assume non_matched_list is the list returned by the get_non_matched_elements function
# and spatial_units is your spatial units data frame
closest_matches_list <- get_closest_matches(non_matched_list, spatial_units)
}

}
