% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{get_column_types}
\alias{get_column_types}
\title{Get column types from metadata}
\usage{
get_column_types(ds)
}
\arguments{
\item{ds}{the dataset}
}
\description{
Get column types from metadata
}
\examples{
\dontrun{
  get_column_types(ds)
}
}
