% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db.R
\name{get_data_types_from_metadata}
\alias{get_data_types_from_metadata}
\title{Get data types for postgres table from metadata}
\usage{
get_data_types_from_metadata(ds, year_name)
}
\arguments{
\item{ds:}{dataset of the pipelin}

\item{year_name:}{name for the year column that depends on the language}
}
\value{
data_types: a tibble with 2 columns: name and data_type
}
\description{
Returns a tibble with 2 columns: the name of the column and the
data_type of the column
This includes also the common columns such as 'uid': the primary key
\code{spatialunit_uid}: the spatial unit foreign key and the \code{year} column
The data_types are postgres data_types that can be used in the postgres
create table statement
}
\examples{
\dontrun{
  get_data_types_from_metadata(ds, "jahr")
}
}
