% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{get_examples}
\alias{get_examples}
\title{get example values for the columns}
\usage{
get_examples(ds, columns)
}
\arguments{
\item{ds}{the dataset: it is expected to have
ds$postgres_export}

\item{columns}{columns of the dataset: colnames(ds$postgres)}
}
\value{
example_values array of example values
}
\description{
get example values for the columns
}
\examples{
\dontrun{
get_examples(ds, columns)
# returns 3 values for each column from the samples
}
}
