% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db_metadata.R
\name{get_metadata_from_postgres}
\alias{get_metadata_from_postgres}
\title{Get metadata from postgres}
\usage{
get_metadata_from_postgres(ds_table_name = NULL)
}
\arguments{
\item{ds_table_name}{name of the table to get metadata for: optional}
}
\value{
list of metadata tables from postgres as tibbles
}
\description{
Read metadata from the postgres instance and return the metadata
of all or a selected table
}
\examples{
\dontrun{
get_metadata_from_postgres()
}
}
