% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_non_matched_elements.R
\name{get_non_matched_elements}
\alias{get_non_matched_elements}
\title{Get Non-Matched Elements for Each Spatial Unit Ontology}
\usage{
get_non_matched_elements(
  spatial_units,
  ds_data,
  spatial_unit_ontologies,
  spatial_units_col
)
}
\arguments{
\item{spatial_units}{A data frame containing spatial unit data.}

\item{ds_data}{A data frame containing the data to be matched.}

\item{spatial_unit_ontologies}{A character vector specifying the spatial unit ontologies to be considered.}

\item{spatial_units_col}{A string specifying the column in the 'spatial_units' data frame to look for matches.}
}
\value{
A named list of data frames. Each list element corresponds to a spatial unit ontology and contains the non-matched elements for that ontology.
The names of the list elements are the spatial unit ontologies.
}
\description{
This function returns a list of data frames containing the non-matched elements for each specified spatial unit ontology.
Non-matched elements are identified by comparing the 'grossregion_kanton' column of the 'ds_data' data frame
with the specified column of the 'spatial_units' data frame for each spatial unit ontology.
}
\examples{
\dontrun{
  # Assume spatial_units and ds_data are your data frames, you're interested in 'Country' and 'Canton', and you want to look for matches in the 'name_de' column
  ontologies <- c("Country", "Canton")
  result <- get_non_matched_elements(spatial_units, ds_data, ontologies, "name_de")
}

}
