% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db.R
\name{get_postgres_data_types}
\alias{get_postgres_data_types}
\title{Get data types for postgres table from metadata}
\usage{
get_postgres_data_types(db, tbl)
}
\arguments{
\item{con:}{db connection}

\item{tbl:}{dataset table to export (ds$postgres_export)}
}
\value{
data_types: a tibble with 2 columns: name and data_type
}
\description{
Returns a tibble with 2 columns: the name of the column and the
data_type of the column
This includes also the common columns such as 'uid': the primary key
\code{spatialunit_uid}: the spatial unit foreign key and the \code{year} column
The data_types are postgres data_types that can be used in the postgres
create table statement
}
\examples{
\dontrun{
get_data_types_from_metadata(DBI::ANSI(), ds$postgres_export)
}
}
