% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db_intial_setup.R
\name{intial_setup_of_metadata_tables}
\alias{intial_setup_of_metadata_tables}
\title{Intitial setup of metadata tables}
\usage{
intial_setup_of_metadata_tables(example_pipeline = "A6")
}
\arguments{
\item{example_pipeline}{example pipeline that is used to derive the structure
of the metadata tables from}
}
\description{
This function was used only once for the inital setup of the tables
}
\examples{
\dontrun{
intial_setup_of_metadata_tables()
}
}
