% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db.R
\name{list_tables_in_statbot_db}
\alias{list_tables_in_statbot_db}
\title{List Postgres table in Statbot DB}
\usage{
list_tables_in_statbot_db()
}
\description{
List Postgres table in Statbot DB
}
\examples{
\dontrun{
list_tables_in_statbot_db()
}
}
