% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db_metadata.R
\name{list_tables_with_metadata_in_statbot_db}
\alias{list_tables_with_metadata_in_statbot_db}
\title{List all tables that have metadata in postgres}
\usage{
list_tables_with_metadata_in_statbot_db()
}
\value{
list of tables with metadata in postgres
}
\description{
List all tables that have metadata in postgres
}
\examples{
\dontrun{
list_tables_with_metadata_in_postgres()
}
}
