% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_dataset_list}
\alias{load_dataset_list}
\title{load dataset list}
\usage{
load_dataset_list()
}
\value{
dataset_list list of datasets in the csv file
}
\description{
loads the list of datasets from a csv file
}
