% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testrun_queries.R
\name{log_question_and_query}
\alias{log_question_and_query}
\title{Writing the sql questinon and query to the output file}
\usage{
log_question_and_query(count, question, query, output_path)
}
\arguments{
\item{count}{the number of the query
(to keep track of how many queries there are)}

\item{question}{the natural language questions that belongs to the query}

\item{query}{the sql query}

\item{output_path}{the file where the query results are logged}
}
\description{
Writing the sql questinon and query to the output file
}
\examples{
\dontrun{
  # command to call. this function
  log_question_and_query(
    1,
    "Wieviele Abstimmungsvorlagen gab es seit dem Jahr 1971?",
    "SELECT COUNT(*) as anzahl_vorlagen_in_2000 FROM abstimmungsvorlagen_seit_1971",
    output_path
  )
}
}
