% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testrun_queries.R
\name{log_result}
\alias{log_result}
\title{Writing the sql result to the output file}
\usage{
log_result(result, output_path)
}
\arguments{
\item{result}{the result of the query}

\item{output_path}{the file where the query results are logged}
}
\description{
Since the reslt is a tibble a different method is used for writing the
result.
}
\examples{
\dontrun{
  # command to call. this function
  log_question_and_query(
    result,
    output_path
  )
}
}
