% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testrun_queries.R
\name{log_time_and_environment}
\alias{log_time_and_environment}
\title{Log environment and time to the output file}
\usage{
log_time_and_environment(pipeline_status, output_path)
}
\arguments{
\item{output_path}{the file where the query results are logged}

\item{pipleine_status}{property status of the dataset class ds}
}
\description{
Log environment and time to the output file
}
\examples{
\dontrun{
log_question_and_query(ds$status, output_path)
}
}
