% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_mappings.R
\name{map_canton_and_country}
\alias{map_canton_and_country}
\title{Map spatial units for the level country and canton}
\usage{
map_canton_and_country(pattern_function)
}
\arguments{
\item{pattern_function}{function to make a pattern for detecting
the spatial unit, default statab_canton_country_pattern
as an example}
}
\value{
spatial_unit_df a tibble with spatial units on the level
of country and canton
}
\description{
Currently the function filters the spatial units for the level of countries
and cantons. It reads in the spatial units stored in postgres from a file.
}
\examples{
\dontrun{
# get a tibble with the spatial units on the levels country and canton
df_spatial <- map_canton_and_country()
}
}
