% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_mappings.R
\name{map_ds_municipalities}
\alias{map_ds_municipalities}
\title{Get spatialunits of municipalities.}
\usage{
map_ds_municipalities(.data, year, canton_abbr, municipality_name)
}
\arguments{
\item{.data}{A tibble containing at least columns with the year of observation, 2 letter canton abbreviations and names of municipalities.}

\item{year}{column containing the year of observation.}

\item{canton_abbr}{column containing the 2 letter abbreviation corresponding to the municipality's canton.}

\item{municipality_name}{column containing the names of municipalities.}
}
\value{
spatial mapping with input columns, plus valid_until, valid_from, spatialunit and name.
}
\description{
This function uses temporal and cantonal information to
unambiguously assign a uid to each municipality. It handles duplicate names in different cantons,
merged, renamed or split municipalities.
}
