% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_mappings.R
\name{map_ds_residential_areas}
\alias{map_ds_residential_areas}
\title{Get spatialunits of residential areas.}
\usage{
map_ds_residential_areas(.data, year, canton_hist_id, residential_area_name)
}
\arguments{
\item{.data}{A tibble containing at least columns with the year of observation, 2 letter canton abbreviations and names of residential areas.}

\item{year}{column containing the year of observation.}

\item{canton_hist_id}{column containing the residential area canton's historical id.}

\item{residential_area_name}{column containing the names of residential areas.}
}
\value{
spatial mapping with input columns, plus valid_until, valid_from, spatialunit and name.
}
\description{
This function uses temporal and cantonal information to
unambiguously assign a uid to each residential area.
}
