% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_mappings.R
\name{map_ds_spatial_units}
\alias{map_ds_spatial_units}
\title{Map spatial units in the observations}
\usage{
map_ds_spatial_units(
  df_spatial,
  spatial_dimensions = c("Canton", "Country"),
  pattern_function = stattab_make_pattern,
  match_function = stattab_find_match
)
}
\arguments{
\item{df_spatial}{tibble that has just one colunm with all the distinct
spatial values of the dataset}

\item{spatial_dimensions}{a vector of values for spatialunit_ontology
default: "Canton" and "Country"}

\item{pattern_function}{function to make a pattern for detecting
the spatial unit, default stattab_make_pattern
as an example}

\item{match_function}{function to detect the pattern in the spatial units
of the dataset, default stattab_find_match
as an example}
}
\value{
df_spatial tibble that contains a mapping of the spatial units
to spatialunit_uid and the name of the spatial unit
}
\description{
This function expects a tibble with the distinct spatial values of the
dataset in this column.
}
\details{
The idea is to find matches for these spatial values in a reduced dataframe
this dataframe is then returned with spatial matches and can be joined
with the dataset in order to transfer the match making to the dataset

In order to keep that function general: the pattern_function and the
match_function can be provided.

The spatial_dimensions should given in order to allow for filtering
}
\examples{
map_ds_spatial_units(df_spatial)
\dontrun{
  # df_spatial should be a tibble with just one column that contains
  # spatial values that are all distinct
  df_spatial <- map_ds_spatial_units(df_spatial)
}
}
