% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_mappings.R
\name{map_spatial}
\alias{map_spatial}
\title{Title}
\usage{
map_spatial(
  term,
  spatial_units_list,
  match_function = stattab_country_canton_find_match
)
}
\arguments{
\item{term}{term that should be matched}

\item{spatial_units_list}{list of spatial units: names are the pattern
used for the matching and values are the spatial
uids that the terms should be matched to}

\item{match_function}{function to detect the pattern in the spatial units
of the dataset, default stattab_country_canton_find_match
as an example}
}
\value{
spatial_code if match was found otherwise blank
}
\description{
Title
}
\examples{
\dontrun{
  # df_spatial should be a tibble with just one column that contains
  # spatial values that are all distinct
  spatialunit_uid <- ("Zug",
                      spatial_units_list
                      stattab_country_canton_find_match)
}
}
