% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_data.json}
\alias{read_data.json}
\title{Method to retrieve a JSON dataset and convert it to a tibble}
\usage{
\method{read_data}{json}(ds)
}
\description{
Method to retrieve a JSON dataset and convert it to a tibble
}
