% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data.rdf}
\alias{read_data.rdf}
\title{Method to retrieve an RDF dataset using an input SELECT query
Execute a SPARQL query and return the result as a tibble.}
\usage{
\method{read_data}{rdf}(ds)
}
\description{
Method to retrieve an RDF dataset using an input SELECT query
Execute a SPARQL query and return the result as a tibble.
}
