% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{read_metadata_tables}
\alias{read_metadata_tables}
\title{Read/Write metadata tables as csv}
\usage{
read_metadata_tables(ds)
}
\arguments{
\item{ds}{the dataset}
}
\value{
metadata as tibbles
}
\description{
Read/Write metadata tables as csv
}
\examples{
\dontrun{
  read_metadata_tables(ds)
}
}
