% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{read_metadata_tables_from_file}
\alias{read_metadata_tables_from_file}
\title{Read metadata for a pipeline from file}
\usage{
read_metadata_tables_from_file(ds)
}
\arguments{
\item{ds}{the dataset}
}
\value{
metadata as tibbles or NULL
}
\description{
Read metadata for a pipeline from file
}
\examples{
\dontrun{
  read_metadata_tables_from_file(ds)
}
}
