% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{read_write_metadata_tables}
\alias{read_write_metadata_tables}
\title{Read/Write metadata tables as csv}
\usage{
read_write_metadata_tables(ds, overwrite = FALSE)
}
\arguments{
\item{ds}{the dataset: it is expected to have
ds$postgres_export
the language is to be expected the language
the ds is in (one language per pipeline)}

\item{overwrite}{default is FALSE, when set to TRUE
the metadata file will be overwritten}
}
\value{
metadata a tibble containing the metadata that
matches what is on file
the function watches for files:
metadata_tables.csv
metadata_table_columns.csv
If they are not there or overwrite is set
to TRUE templates for these files will be
set up. Otherwise the files will be read
and reported back
}
\description{
Read/Write metadata tables as csv
}
\examples{
\dontrun{
  # command to call when the metadata should be read or
  # created: it will be created if it does not exist yet
  read_write_metadata_tables(ds)

  # command to call when the metadata should be overwritten
  read_write_metadata_tables(ds, overwrite = TRUE)
}
}
