% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db.R
\name{run_postgres_query}
\alias{run_postgres_query}
\title{Run postgres query on table in schema}
\usage{
run_postgres_query(table_name, query)
}
\arguments{
\item{table_name}{table_name in the query}

\item{query}{query on the table}
}
\description{
Add schema to query
}
\examples{
\dontrun{
run_postgres_query("SELECT count(*) FROM metadata.tables")
}
}
