% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_table.R
\name{rw_metadata_table}
\alias{rw_metadata_table}
\title{Read/Write metadata table as csv}
\usage{
rw_metadata_table(ds, overwrite = FALSE)
}
\arguments{
\item{ds}{the dataset: it is expected to have
ds$postgres_export
the language is to be expected the language
the ds is in (one language per pipeline)}

\item{overwrite}{default is FALSE, when set to TRUE
the metadata file will be overwritten}
}
\value{
metadata a tibble containing the metadata that
matches what is on file
}
\description{
Read/Write metadata table as csv
}
\examples{
\dontrun{
  # command to call when the metadata should be read or
  # when metadata do not exist yet
  rw_metadata_table(ds)

  # command to call when the metadata should be overwritten
  rw_metadata_table(ds, overwrite = TRUE)
}
}
