% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_mappings.R
\name{spatial_mapping_country}
\alias{spatial_mapping_country}
\title{Get Spatial Unit for Switzerland}
\usage{
spatial_mapping_country()
}
\value{
spatialunit_uid_ch spatialunit_uid for Switzerland
}
\description{
This is a very simple function to provide the spatialunit_uid for Switzerland
The spatial unit has to be added to tables that don't come with a spatial
unit
}
\examples{
spatial_mapping_country()
\dontrun{
  spatialunit_uid_ch <- spatial_mapping_country()
}
}
