% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_mappings.R
\name{stattab_find_match}
\alias{stattab_find_match}
\title{Stattab find match}
\usage{
stattab_find_match(spacial_term, spatial_units_list)
}
\arguments{
\item{spacial_term}{spatial term that should be matched}

\item{spatial_units_list}{list of spatialunit uid
the list is expected to be named and the names
are the pattern used for the matching}
}
\value{
match or blank if no match was found
}
\description{
The functions finds a map for a spatial term considering a given map
of pattern and matching targets
}
\examples{
\dontrun{
.  # Assume non_matched_list is the list returned by the get_non_matched_elements function
  # and spatial_units is your spatial units data frame
  match <- stattab_find_match("Basel-Stadt",
                              list("basel_stadt,bale_ville,basilea_citta" = "12_A.ADM1"))
}
}
