% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_mappings.R
\name{stattab_make_pattern}
\alias{stattab_make_pattern}
\title{Stattab pattern making}
\usage{
stattab_make_pattern(name, name_de, name_fr, name_it)
}
\arguments{
\item{name}{name of the spatial unit}

\item{name_de}{german translation}

\item{name_fr}{french translation}

\item{name_it}{italian translation}
}
\value{
pattern that is used for the match making
}
\description{
The function makes a pattern to detect a spatial unit
by cleaning all translations and combining them into
a comma separated string. This functions has to correspond to
the pattern detection function: stattab_find_match
}
\examples{
\dontrun{
  # Assume non_matched_list is the list returned by the get_non_matched_elements function
  # and spatial_units is your spatial units data frame
  stattab_make_pattern("Canton of Basel-Stadt","Kanton Basel-Stadt",
                       "Canton de Bâle-Ville","Cantone di Basilea Città")
  # returns: "basel_stadt,bale_ville,basilea_citta"
}

}
