% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testrun_queries.R
\name{testrun_queries}
\alias{testrun_queries}
\title{Testrun queries}
\usage{
testrun_queries(df, dir, table_name)
}
\arguments{
\item{df}{tibble to turn into an sql table}

\item{dir}{directory to write the output to}

\item{table_name.}{name of the table in sql database}
}
\description{
the database will just have two table, one that is defined
by the provided tibble and the
the queries are expected to be read from a file
\code{queries.sql} in the directory \code{dir}
}
\examples{
#' \dontrun{
  # command to call. this function
  testrun_queries(ds$data, ds$dir, ds$name)
  # example input file `queries.sql`
  -- SELECT COUNT(*) from -- Wieviele Abstimmungsvorlagen gab es seit dem Jahr 1971?
  SELECT COUNT(*) as anzahl_vorlagen_in_2000
.  FROM abstimmungsvorlagen_seit_1971 as T
.  JOIN spatial_unit as S on T.spatialunit_uid = S.spatialunit_uid
.  WHERE S.name_de='Schweiz' AND S.country=TRUE
}
#' #' \dontrun{
  testrun_queries(ds$data, ds$dir, ds$name)
}
}
