% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db_metadata.R
\name{update_metadata_in_postgres}
\alias{update_metadata_in_postgres}
\title{Update metadata for a pipeline in postgres}
\usage{
update_metadata_in_postgres(ds)
}
\arguments{
\item{ds}{dataset pipeline class}
}
\value{
list of metadata tables from postgres as tibbles
}
\description{
Update the metadata in postgres by deleting and
inserting it again
}
\examples{
\dontrun{
update_metadata_in_postgres(ds)
}
}
