% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{update_pipeline_last_run_date}
\alias{update_pipeline_last_run_date}
\title{Update pipeline last run date in the table metadata on file}
\usage{
update_pipeline_last_run_date(ds)
}
\arguments{
\item{ds}{the dataset}
}
\description{
Update pipeline last run date in the table metadata on file
}
\examples{
\dontrun{
  update_pipeline_last_run_date(ds)
}
}
