% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db_intial_setup.R
\name{upload_spatial_units_to_postgres}
\alias{upload_spatial_units_to_postgres}
\title{Load spatial units into a spatial_unit_table}
\usage{
upload_spatial_units_to_postgres()
}
\description{
The table has already been set up at the remote postgres
instance. Just the data was loaded from file.
This function was only used once during the setup phase.
}
\examples{
\dontrun{
upload_spatial_units_to_postgres()
}
}
