% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db.R
\name{create_postgres_table}
\alias{create_postgres_table}
\title{Create Postgres table}
\usage{
create_postgres_table(ds, dry_run = FALSE)
}
\arguments{
\item{dry_run}{bool: whether this}
}
\value{
field_types: vector of field types, in case it was called with dry_run
}
\description{
Create Postgres table
}
\examples{
\dontrun{
  create_postgres_table(ds, dry_run = TRUE)
}
}
