% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_sample.R
\name{dataset_sample}
\alias{dataset_sample}
\title{Write sample for the dataset}
\usage{
dataset_sample(ds, sample_size = 40)
}
\arguments{
\item{ds}{dataset pipeline class}

\item{sample_size}{default 40}
}
\description{
The sample contains random rows
The number of the rows defaults to 40
but can be specified
}
\examples{
\dontrun{
dataset_sample(ds)
}
}
