% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db_metadata.R
\name{delete_metadata_from_postgres}
\alias{delete_metadata_from_postgres}
\title{Delete metadata for a pipeline in postgres}
\usage{
delete_metadata_from_postgres(table_name)
}
\arguments{
\item{table_name}{name of the table to delete metadata for}
}
\description{
The pipeline dataset is supposed to be created when this function
is called. All metadata for the pipeline is deleted in the postgres
instance. This function can be used when the metadata should be updated
to remove the old entries and the readd the updated metadata
}
\examples{
\dontrun{
delete_metadata_from_postgres(table_name)
}
}
