% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db_metadata.R
\name{get_metadata_from_postgres}
\alias{get_metadata_from_postgres}
\title{Get metadata from postgres}
\usage{
get_metadata_from_postgres(ds = NULL)
}
\arguments{
\item{ds}{dataset pipeline class}
}
\value{
list of metadata tables from postgres as tibbles
}
\description{
Read metadata from the postgres instance and return the metadata
tables: the full tables are returned.
}
\examples{
\dontrun{
get_metadata_from_postgres()
}
}
