% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_db.R
\name{postgres_db_connect}
\alias{postgres_db_connect}
\title{Establish db connection}
\usage{
postgres_db_connect()
}
\value{
db_connection: list of a running db connection and a schema
}
\description{
Establish db connection
}
\examples{
\dontrun{
  postgres_db_connect()
}
}
