% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testrun_queries.R
\name{prepare_test_db}
\alias{prepare_test_db}
\title{Preparing the database table for the query test runs}
\usage{
prepare_test_db(db_con, df, table_name)
}
\arguments{
\item{db_con}{sql connection}

\item{df}{tibble that gets loaded as a sql table}

\item{table_name}{name of the table}
}
\description{
The spatial_unit tables is needed for the spatial joins.
}
\examples{
\dontrun{
  # command to call. this function
  prepare_db(
    db_con,
    df,
    'abstimmungsvorlagen_seit_1971'
  )
}
}
